+++ packages/view/src/i18n/en-US.js
@@ -1,0 +1,8 @@
+export default {
+  types: {
+    outline: 'outline',
+    filled: 'filled',
+  },
+  search: 'enter icon name to search',
+  copied: 'copied',
+};