+++ packages/view/src/i18n/zh-CN.js
@@ -1,0 +1,9 @@
+export default {
+  types: {
+    outline: '描边风格',
+    filled: '填充风格',
+  },
+  search: '输入图标名称搜索',
+  copied: '复制成功',
+
+};