+++ packages/vue/src/svg-sprite/svg-sprite.tsx
@@ -11,7 +11,7 @@ const { classPrefix } = ConfigContext;
 
 const tName = `${classPrefix}-icon`;
 
-const CDN_ICONFONT_URL = 'https://tdesign.gtimg.com/icon/0.1.4/fonts/index.js';
+const CDN_ICONFONT_URL = 'https://tdesign.gtimg.com/icon/0.2.0/fonts/index.js';
 
 export default Vue.extend({
   name: 'Icon',