+++ packages/svg/CHANGELOG.md
@@ -1,5 +1,11 @@
 # tdesign-icons-svg
 
+## 0.2.0
+
+### Minor Changes
+
+- feat: increase 900 + new icons
+
 ## 0.1.8
 
 ### Patch Changes