+++ packages/vue-next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.2.0
+
+### Minor Changes
+
+- feat: increase 900 + new icons
+
 ## 0.1.12
 
 ### Patch Changes